Move PlayerInterface to models/game/interfaces/characters
