+++ frontend/models/game/Game.php
@@ -12,12 +12,11 @@ namespace frontend\models\game;
 use frontend\exceptions\AlreadyStartedGameException;
 use frontend\exceptions\CannotStartWithoutCharacterException;
 use frontend\exceptions\FinishedGameException;
-use frontend\exceptions\ImpossibleCaseException;
 use frontend\exceptions\NotStartedGameException;
 use frontend\models\game\base\BeeInterface;
 use frontend\models\game\base\CharacterPoolInterface;
 use frontend\models\game\base\HoneyPoolInterface;
-use frontend\models\game\characters\PlayerInterface;
+use frontend\models\game\interfaces\characters\PlayerInterface;
 
 class Game implements GameInterface
 {
